requirements.txt
gradio>=5.50.0,<6
pandas>=2.2
numpy>=1.26
pyyaml>=6.0
python-dateutil>=2.9
requests>=2.32
faster-whisper>=1.1.0
piper-tts>=1.3.0
